CachedI18NMessageBundle: add a comment about a possible timeout parameter
Right now, the result of the FutureTask is grabbed using .get(). And .get() will
return when the underlying thread is complete.

Potentially, it means a thread may sleep forever. But if we decide on using a
timeout, we have to decide on a value for this timeout. The risk always exists
of making it either too short or too long. And making it configurable may be
overkill.