New interface MessageSourceProvider
This helps provide a MessageSource for a given locale.

Future plans involve querying an ordered list of such providers from a
MessageBundle. In turn, this means the result from invoking this interface's
only method may vary according to implementations.

The latter point gives the future liberty for implementations to implement a
timeout/expiration policy of sources as well. Also, fixed providers can choose
to always return the same result for a given locale.