LocaleUtils: comment potential bug in locale parsing
Any input string with more than two underscores (therefore with four parts or
more when splitting it) is theoretically illegal.

But then we don't check the validity of either the language, country or variant
al all; is it allowed that a locale variant have an underscore? Right now, it is
surmised that this is illegal.