Eat own dog's food and puke -- ensure bundles are really loaded
Here it was chosen to use static initialization blocks to load all services.
Unfortunately, there is no guarantee as to what block is executed first.

Ensure correct initialization by making MessageBundles a "private singleton",
and synchronize on the class to ensure that the singleton instance is fully
initialized before we attempt to return a bundle.