PropertiesBundle: modify comment now that it is know what is happening
Static initializations are done in order. And when initializing bundles, one may
want to instantiate an instance of this class...

Which means, if the bundle is attempted at initialization _before_ SUFFIX is
initialized, SUFFIX will be null!

Pre-declare the bundle in LoadingMessageSourceProvider with the same comment as
a warning.

Finally, also comment on the INSTANCE variable static but not final in
MessageBundles.