PropertiesMessageSourceTest: add a test to confirm a behaviour of Properties
I wasn't sure what the method retrieving a string array did; now I know: it
expects a semicolon as a separator, and if found, it splits into a string array.

Confirm, with this test, that such a property is correctly understood by
.stringPropertyNames().

Good thing that there was _at last_ a test case I could rely upon, the Javadoc
didn't help at all here!

Source: http://stackoverflow.com/questions/17170248/enum-relating-to-a-string-array-that-was-read-from-file